requirements.txt
Django>=3.2.6,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<2.9.9
Pillow>=8.3.1,<8.3.9
flake8>=3.6.0,<3.7.0